+++ src/proto/mod.rs
@@ -1,1 +1,3 @@
+pub mod errors;
 pub mod h1;
+pub mod util;
